fix: remove old PreviewUnboundImage and unnnecessary special casing
