+++ packages/core/src/utils/sanitizeNodeProps.ts
@@ -4,7 +4,14 @@ import type { PrimitiveValue } from '../types';
 
 const stylesToKeep = ['cfImageAsset'];
 const stylesToRemove = CF_STYLE_ATTRIBUTES.filter((style) => !stylesToKeep.includes(style));
-const propsToRemove = ['cfHyperlink', 'cfOpenInNewTab', 'cfSsrClassName'];
+// cfWrapColumns & cfWrapColumnsCount are no real style attributes as they are handled on the editor side
+const propsToRemove = [
+  'cfHyperlink',
+  'cfOpenInNewTab',
+  'cfSsrClassName',
+  'cfWrapColumns',
+  'cfWrapColumnsCount',
+];
 
 export const sanitizeNodeProps = (nodeProps: Record<PropertyKey, PrimitiveValue>) => {
   return omit(nodeProps, stylesToRemove, propsToRemove);
