+++ packages/experience-builder-sdk/src/core/styles/createStylesheetsForBuiltInStyles.ts
@@ -4,14 +4,12 @@ import {
   flattenDesignTokenRegistry,
   maybePopulateDesignTokenValue,
   stringifyCssProperties,
-  toMediaQuery,
   buildCfStyles,
 } from '@contentful/experiences-core';
 import { Breakpoint, ComponentTreeNode, PrimitiveValue } from '@contentful/experiences-core/types';
 import md5 from 'md5';
-import { useMemo } from 'react';
 
-type ResolvedStylesheetData = Array<{
+export type ResolvedStylesheetData = Array<{
   className: string;
   breakpointCondition: string;
   css: string;
@@ -101,72 +99,4 @@ export const createStylesheetsForBuiltInStyles = ({
   }
 
   return result;
-};
-
-/**
- * Takes the CSS code for each breakpoint and merges them into a single CSS string.
- * It will wrap each breakpoint's CSS code in a media query (exception: default breakpoint with '*').
- *
- * **Example Input:**
- * ```
- * [
- *  { className: 'cfstyles-123', breakpointCondition: '*', css: 'color:red;' },
- *  { className: 'cfstyles-456', breakpointCondition: '<768px', css: 'color:blue;' },
- * ]
- * ```
- *
- * **Example Output:**
- * ```
- * '.cfstyles-123{color:red;}@media(max-width:768px){.cfstyles-456{color:blue;}}'
- * ```
- */
-export const convertResolvedDesignValuesToMediaQuery = (stylesheetData: ResolvedStylesheetData) => {
-  const stylesheet = stylesheetData.reduce(
-    (acc, { breakpointCondition, className, css }) => {
-      if (acc.classNames.includes(className)) {
-        return acc;
-      }
-
-      const mediaQueryCss = toMediaQuery({
-        condition: breakpointCondition,
-        cssByClassName: { [className]: css },
-      });
-      return {
-        classNames: [...acc.classNames, className],
-        css: `${acc.css}${mediaQueryCss}`,
-      };
-    },
-    {
-      classNames: [] as string[],
-      css: '',
-    },
-  );
-
-  return {
-    css: stylesheet.css,
-    className: stylesheet.classNames.join(' '),
-  };
-};
-
-export const useMediaQuery = ({
-  designPropertiesByBreakpoint,
-  breakpoints,
-  node,
-  patternRootNodeIdsChain,
-}: {
-  designPropertiesByBreakpoint: Record<string, Record<string, any>>;
-  breakpoints: Breakpoint[];
-  node: ComponentTreeNode;
-  patternRootNodeIdsChain?: string;
-}) => {
-  return useMemo(() => {
-    const stylesheetData = createStylesheetsForBuiltInStyles({
-      designPropertiesByBreakpoint,
-      breakpoints,
-      node,
-      patternRootNodeIdsChain,
-    });
-
-    return convertResolvedDesignValuesToMediaQuery(stylesheetData);
-  }, [designPropertiesByBreakpoint, breakpoints, node, patternRootNodeIdsChain]);
 };
